refactors attribute related methods to AttributeProcessor class
